Create deepcopy of dictionary before deleting from it
If no deepcopy of the `grid_mean_predictor` is created and it was given as a default argument, calling `.pop('type)` on it will delete the `type` key in the default argument as well. When the function is called the next time (for example in Table.populate()), it will error out because the `type` key doesn't exist. One would need to restart the kernel and re-import the function for the full dictionary to be there again. 
The solution here is to always create a deepcopy before deleting things out of a function argument